Add benchstat on cobra
