+++ new/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/vitessio/vitess-tester/new/cmd"
+
+func main() {
+	cmd.Execute()
+}
